refactor: move hasAnyMark to test helpers
Test-only utility was defined in the production documentHelpers
module. Move it to @tests/helpers/helpers.js so it doesn't ship
in the production bundle.
